feat/first part

+++ app/Http/Controllers/API/Admin/CategoryController.php
@@ -5,9 +5,13 @@ namespace App\Http\Controllers\API\Admin;
 use App\Category;
 use App\Http\Controllers\Controller;
 use App\Http\Requests\CategoryRequest;
+use App\Image;
 use Illuminate\Database\Eloquent\ModelNotFoundException;
 use Illuminate\Http\Request;
 use Illuminate\Http\JsonResponse;
+use Illuminate\Support\Facades\DB;
+use Illuminate\Support\Facades\Storage;
+
 
 class CategoryController extends Controller
 {
@@ -19,8 +23,10 @@ class CategoryController extends Controller
     public function index()
     {
         try {
-            $categories = Category::orderBy('slug')->paginate(10);
-            return response()->json($categories, JsonResponse::HTTP_OK);
+            $categories = Category::orderBy('slug')->with(['image' => function ($query) {
+                $query->select('*', DB::raw("CONCAT('" . Storage::disk('public')->url('images/categories/') . "', file_name) AS url"));
+            }])->paginate(10);
+            return response()->json(['status' => 'success', 'message' => 'Categories indexed successfully!', 'categories' => $categories], JsonResponse::HTTP_OK);
 
         } catch (\Exception $ex) {
             return response()->json(['message' => $ex->getMessage()], JsonResponse::HTTP_CONFLICT);
@@ -30,7 +36,7 @@ class CategoryController extends Controller
     /**
      * Store a newly created resource in storage.
      *
-     * @param  \Illuminate\Http\Request  $request
+     * @param  \Illuminate\Http\Request $request
      * @return \Illuminate\Http\Response
      */
     public function store(CategoryRequest $request)
@@ -40,7 +46,16 @@ class CategoryController extends Controller
             $category = new \App\Category();
             $category->fill($data);
             $category->save();
-            return response()->json( ['message' => 'successfully added'],JsonResponse::HTTP_OK);
+            if ($request->hasFile('image')) {
+                $file = $request->file('image');
+                $filename = time() . '.' . $file->getClientOriginalExtension();
+                Storage::disk('public')->putFileAs('categories', $file, $filename);
+                $image = new Image();
+                $image->file_name = $filename;
+                $category->image()->save($image);
+            }
+
+            return response()->json(['status' => 'success', 'message' => 'Categories add successfully!'], JsonResponse::HTTP_OK);
 
         } catch (\Exception $ex) {
             return response()->json(['message' => $ex->getMessage()], JsonResponse::HTTP_CONFLICT);
@@ -50,17 +65,18 @@ class CategoryController extends Controller
     /**
      * Display the specified resource.
      *
-     * @param  int  $id
+     * @param  int $id
      * @return \Illuminate\Http\Response
      */
-    public function show($id)
+    public function show($slug)
     {
         try {
-            $category = Category::findOrFail($id);
-            return response()->json($category, JsonResponse::HTTP_OK);
-
-        } catch (ModelNotFoundException $e) {
-            return response()->json(['message' => 'Model dont\t existe'], JsonResponse::HTTP_CONFLICT);
+            $category = Category::where('slug', $slug)->first();
+            if ($category) {
+                return response()->json(['status' => 'success', 'message' => 'Category retrieved successfully!', 'category' => $category], JsonResponse::HTTP_OK);
+            } else {
+                return response()->json(['message' => 'Model dont existe'], JsonResponse::HTTP_CONFLICT);
+            }
 
         } catch (\Exception $ex) {
             return response()->json(['message' => $ex->getMessage()], JsonResponse::HTTP_CONFLICT);
@@ -70,21 +86,34 @@ class CategoryController extends Controller
     /**
      * Update the specified resource in storage.
      *
-     * @param  \Illuminate\Http\Request  $request
-     * @param  int  $id
+     * @param  \Illuminate\Http\Request $request
+     * @param  int $id
      * @return \Illuminate\Http\Response
      */
-    public function update(Request $request, $id)
+    public function update(CategoryRequest $request, $slug)
     {
         try {
             $data = $request->all();
-            $category = Category::findOrFail($id);
-            $category->update($data);
-            return response()->json(['message' => 'successfully updated'], JsonResponse::HTTP_OK);
+            $category = Category::where('slug', $slug)->first();
+            if ($category) {
+                $category->update($data);
 
-        }catch (ModelNotFoundException $e) {
-            return response()->json(['message' => 'Model dont\t existe'], JsonResponse::HTTP_CONFLICT);
+                if ($request->hasFile('image')) {
+                    Storage::disk('public')->delete('categories/' . $category->image->file_name);
+                    $category->image()->delete();
+                    $file = $request->file('image');
+                    $filename = time() . '.' . $file->getClientOriginalExtension();
+                    Storage::disk('public')->putFileAs('categories', $file, $filename);
+                    $image = new Image();
+                    $image->file_name = $filename;
+                    $category->image()->save($image);
+                }
 
+                return response()->json(['status' => 'success', 'message' => 'Category updated successfully!', 'category' => $category], JsonResponse::HTTP_OK);
+
+            } else {
+                return response()->json(['message' => 'Model dont existe'], JsonResponse::HTTP_CONFLICT);
+            }
         } catch (\Exception $ex) {
             return response()->json(['message' => $ex->getMessage()], JsonResponse::HTTP_CONFLICT);
         }
@@ -93,17 +122,19 @@ class CategoryController extends Controller
     /**
      * Remove the specified resource from storage.
      *
-     * @param  int  $id
+     * @param  int $id
      * @return \Illuminate\Http\Response
      */
-    public function destroy($id)
+    public function destroy($slug)
     {
         try {
-            $category = Category::findOrFail($id);
-            $category->delete();
-            return response()->json(['message' => 'successfully deleted'], JsonResponse::HTTP_OK);
-        }catch (ModelNotFoundException $e) {
-            return response()->json(['message' => 'Model dont\t existe'], JsonResponse::HTTP_CONFLICT);
+            $category = Category::where('slug', $slug)->first();
+            if ($category) {
+                $category->delete();
+                return response()->json(['status' => 'success', 'message' => 'Category deleted successfully!'], JsonResponse::HTTP_OK);
+            } else {
+                return response()->json(['message' => 'Model dont existe'], JsonResponse::HTTP_CONFLICT);
+            }
 
         } catch (\Exception $ex) {
             return response()->json(['message' => $ex->getMessage()], JsonResponse::HTTP_CONFLICT);
