+++ app/Http/Controllers/API/Admin/CourseController.php
@@ -5,9 +5,12 @@ namespace App\Http\Controllers\API\Admin;
 use App\Course;
 use App\Http\Controllers\Controller;
 use App\Http\Requests\CourseRequest;
+use App\Image;
 use Illuminate\Database\Eloquent\ModelNotFoundException;
 use Illuminate\Http\Request;
 use Illuminate\Http\JsonResponse;
+use Illuminate\Support\Facades\DB;
+use Illuminate\Support\Facades\Storage;
 
 class CourseController extends Controller
 {
@@ -19,8 +22,16 @@ class CourseController extends Controller
     public function index()
     {
         try {
-            $courses = Course::orderBy('slug')->paginate(10);
-            return response()->json($courses, JsonResponse::HTTP_OK);
+            $courses = Course::orderBy('slug')->with(['image' => function ($query) {
+                $query->select('*', DB::raw("CONCAT('" . Storage::disk('public')->url('images/courses/') . "', file_name) AS url"));
+            }])->with(['category' => function ($query) {
+                $query->with(['image' => function ($query) {
+                    $query->select('*', DB::raw("CONCAT('" . Storage::disk('public')->url('images/categories/') . "', file_name) AS url"));
+                }]);
+            }])
+                ->paginate(10);
+
+            return response()->json(['status' => 'success', 'message' => 'Courses indexed successfully!', 'courses' => $courses], JsonResponse::HTTP_OK);
 
         } catch (\Exception $ex) {
             return response()->json(['message' => $ex->getMessage()], JsonResponse::HTTP_CONFLICT);
@@ -30,7 +41,7 @@ class CourseController extends Controller
     /**
      * Store a newly created resource in storage.
      *
-     * @param  \Illuminate\Http\Request  $request
+     * @param  \Illuminate\Http\Request $request
      * @return \Illuminate\Http\Response
      */
     public function store(CourseRequest $request)
@@ -40,7 +51,17 @@ class CourseController extends Controller
             $course = new \App\Course();
             $course->fill($data);
             $course->save();
-            return response()->json( ['message' => 'successfully added'],JsonResponse::HTTP_OK);
+
+            if ($request->hasFile('image')) {
+                $file = $request->file('image');
+                $filename = time() . '.' . $file->getClientOriginalExtension();
+                Storage::disk('public')->putFileAs('courses', $file, $filename);
+                $image = new Image();
+                $image->file_name = $filename;
+                $course->image()->save($image);
+            }
+
+            return response()->json(['status' => 'success', 'message' => 'Courses add successfully!'], JsonResponse::HTTP_OK);
 
         } catch (\Exception $ex) {
             return response()->json(['message' => $ex->getMessage()], JsonResponse::HTTP_CONFLICT);
@@ -50,17 +71,19 @@ class CourseController extends Controller
     /**
      * Display the specified resource.
      *
-     * @param  int  $id
+     * @param  int $id
      * @return \Illuminate\Http\Response
      */
-    public function show($id)
+    public function show($slug)
     {
         try {
-            $course = Course::findOrFail($id);
-            return response()->json($course, JsonResponse::HTTP_OK);
+            $course = Course::where('slug', $slug)->first();
 
-        } catch (ModelNotFoundException $e) {
-            return response()->json(['message' => 'Model dont\t existe'], JsonResponse::HTTP_CONFLICT);
+            if ($course) {
+                return response()->json(['status' => 'success', 'message' => 'Course retrieved successfully!', 'course' => $course], JsonResponse::HTTP_OK);
+            } else {
+                return response()->json(['message' => 'Model dont existe'], JsonResponse::HTTP_CONFLICT);
+            }
 
         } catch (\Exception $ex) {
             return response()->json(['message' => $ex->getMessage()], JsonResponse::HTTP_CONFLICT);
@@ -70,21 +93,35 @@ class CourseController extends Controller
     /**
      * Update the specified resource in storage.
      *
-     * @param  \Illuminate\Http\Request  $request
-     * @param  int  $id
+     * @param  \Illuminate\Http\Request $request
+     * @param  int $id
      * @return \Illuminate\Http\Response
      */
-    public function update(Request $request, $id)
+    public function update(Request $request, $slug)
     {
         try {
             $data = $request->all();
-            $course = Course::findOrFail($id);
+            $course = Course::where('slug', $slug)->first();
             $course->update($data);
-            return response()->json(['message' => 'successfully updated'], JsonResponse::HTTP_OK);
+            if ($course) {
+                $course->update($data);
 
-        } catch (ModelNotFoundException $e) {
-            return response()->json(['message' => 'Model dont\t existe'], JsonResponse::HTTP_CONFLICT);
+                if ($request->hasFile('image')) {
+                    Storage::disk('public')->delete('courses/' . $course->image->file_name);
+                    $course->image()->delete();
+                    $file = $request->file('image');
+                    $filename = time() . '.' . $file->getClientOriginalExtension();
+                    Storage::disk('public')->putFileAs('courses', $file, $filename);
+                    $image = new Image();
+                    $image->file_name = $filename;
+                    $course->image()->save($image);
+                }
 
+                return response()->json(['status' => 'success', 'message' => 'Course updated successfully!', 'course' => $course], JsonResponse::HTTP_OK);
+
+            } else {
+                return response()->json(['message' => 'Model dont existe'], JsonResponse::HTTP_CONFLICT);
+            }
         } catch (\Exception $ex) {
             return response()->json(['message' => $ex->getMessage()], JsonResponse::HTTP_CONFLICT);
         }
@@ -93,19 +130,19 @@ class CourseController extends Controller
     /**
      * Remove the specified resource from storage.
      *
-     * @param  int  $id
+     * @param  int $id
      * @return \Illuminate\Http\Response
      */
     public function destroy($id)
     {
         try {
-            $course = Course::findOrFail($id);
-            $course->delete();
-            return response()->json(['message' => 'successfully deleted'], JsonResponse::HTTP_OK);
-
-        } catch (ModelNotFoundException $e) {
-            return response()->json(['message' => 'Model dont\t existe'], JsonResponse::HTTP_CONFLICT);
-
+            $course = Course::where('slug', $id)->first();
+            if ($course) {
+                $course->delete();
+                return response()->json(['status' => 'success', 'message' => 'Course deleted successfully!'], JsonResponse::HTTP_OK);
+            } else {
+                return response()->json(['message' => 'Model dont existe'], JsonResponse::HTTP_CONFLICT);
+            }
         } catch (\Exception $ex) {
             return response()->json(['message' => $ex->getMessage()], JsonResponse::HTTP_CONFLICT);
         }
